Seperated index page from App
